refactor: split exec operation payload codec
